+++ src/backend/marsha/core/models/video.py
@@ -216,6 +216,15 @@ class Video(BaseModel):
             result = _("{:s} [deleted]").format(result)
         return result
 
+    @property
+    def is_ready_to_play(self):
+        """Whether the video is ready to play (ie) has been sucessfully uploaded.
+
+        The value of this field seems to be trivially derived from the value of the
+        `uploaded_on` field but it is necessary for conveniency and clarity in the client.
+        """
+        return self.uploaded_on is not None
+
     def get_source_s3_key(self, stamp=None):
         """Compute the S3 key in the source bucket (resource ID + ID of the video + version stamp).
 
@@ -276,6 +285,15 @@ class BaseTrack(BaseModel):
         """Options for the ``BaseTrack`` model."""
 
         abstract = True
+
+    @property
+    def is_ready_to_play(self):
+        """Whether the track is ready to play (ie) has been sucessfully uploaded.
+
+        The value of this field seems to be trivially derived from the value of the
+        `uploaded_on` field but it is necessary for conveniency and clarity in the client.
+        """
+        return self.uploaded_on is not None
 
 
 class AudioTrack(BaseTrack):